AB-1633 Fix lint, rename components
