+++ src/App.tsx
@@ -5,8 +5,9 @@ import { CreateWallet } from './routes/CreateWallet/CreateWallet';
 import { Home } from './routes/Home';
 import { Login } from './routes/Login/Login';
 import { Network } from './routes/Network';
-import { FungibleTokenInfo, FungibleTokenList } from './routes/UnitList/FungibleTokenList';
-import { UnitList } from './routes/UnitList/UnitList';
+import { AggregatedTokenList } from './routes/TokenList/Fungible/AggregatedTokenList';
+import { TokenDetails } from './routes/TokenList/Fungible/TokenDetails';
+import { TokenList } from './routes/TokenList/TokenList';
 
 export function App(): ReactElement {
   const [, setIsNetworkError] = useState<boolean>(false);
@@ -27,10 +28,10 @@ export function App(): ReactElement {
         <Routes>
           <Route path="/" element={<Home />}>
             <Route path="network" element={<Network />} />
-            <Route path="units" element={<UnitList />}>
-              <Route path="fungible" element={<FungibleTokenList />} />
+            <Route path="units" element={<TokenList />}>
+              <Route path="fungible" element={<AggregatedTokenList />} />
             </Route>
-            <Route path="units/fungible/:id" element={<FungibleTokenInfo />} />
+            <Route path="units/fungible/:id" element={<TokenDetails />} />
             <Route path="" element={<Navigate to="/units/fungible" replace={true} />} />
           </Route>
           <Route path="/create-wallet" element={<CreateWallet />} />
