+++ src/routes/TokenList/TokenList.scss
