+++ vite.config.ts
@@ -1,16 +1,16 @@
-import react from "@vitejs/plugin-react";
-import { defineConfig } from "vite";
-import svgr from "vite-plugin-svgr";
+import react from '@vitejs/plugin-react';
+import { defineConfig } from 'vite';
+import svgr from 'vite-plugin-svgr';
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react(), svgr()],
   build: {
-    outDir: "build",
+    outDir: 'build',
     chunkSizeWarningLimit: 2000,
   },
   define: {
-    global: "globalThis",
+    global: 'globalThis',
   },
   server: {
     port: 3000,
